renaming; training runs

+++ models/2d_model/utils.py
@@ -198,7 +198,7 @@ class CustomAffs(BatchFilter):
         logger.debug("computing ground-truth affinities from labels")
 
         affinities = seg_to_affgraph(
-            batch.arrays[self.labels].data.astype(np.int32), self.affinity_neighborhood
+            batch.arrays[self.labels].data.astype(np.uint64), self.affinity_neighborhood
         ).astype(self.dtype)
 
         # crop affinities to requested ROI
